move function to own file saves 40k
