+++ pages/_document.tsx
@@ -1,7 +1,7 @@
 import Document, { DocumentContext, Head, Main, Html, NextScript } from "next/document"
 import * as React from "react"
 import { AppRegistry, I18nManager } from "react-native-web"
-import { setDimensionsForScreen } from "src/layout/ScreenSize"
+import { setDimensionsForScreen } from "src/layout/setDimensionsForScreen"
 import { getSentry } from "src/utils/sentry"
 
 interface NextReq {
@@ -43,31 +43,27 @@ export default class MyDocument extends Document<Props> {
   render() {
     const { locale } = this.props
     return (
-      <Html lang={locale} style={{ height: '100%', width: '100%' }}>
+      <Html lang={locale} style={{ height: "100%", width: "100%" }}>
         <Head>
           <link key="favicon" rel="shortcut icon" type="image/x-icon" href="/favicon.ico" />
-          <link
-            rel="stylesheet"
-            href="/fonts/JostFont.css"
-            type="text/css"
-          />
-          <link
-            rel="stylesheet"
-            href="/fonts/EBGaramondFont.css"
-            type="text/css"
-          />
-
+          <link rel="stylesheet" href="/fonts/JostFont.css" type="text/css" />
+          <link rel="stylesheet" href="/fonts/EBGaramondFont.css" type="text/css" />
         </Head>
         <body>
           <Main />
           <NextScript />
         </body>
-        <script type="text/javascript" dangerouslySetInnerHTML={{__html: `
+        <script
+          type="text/javascript"
+          dangerouslySetInnerHTML={{
+            __html: `
                   (function(e,t,o,n,p,r,i){e.visitorGlobalObjectAlias=n;e[e.visitorGlobalObjectAlias]=e[e.visitorGlobalObjectAlias]||function(){(e[e.visitorGlobalObjectAlias].q=e[e.visitorGlobalObjectAlias].q||[]).push(arguments)};e[e.visitorGlobalObjectAlias].l=(new Date).getTime();r=t.createElement("script");r.src=o;r.async=true;i=t.getElementsByTagName("script")[0];i.parentNode.insertBefore(r,i)})(window,document,"https://diffuser-cdn.app-us1.com/diffuser/diffuser.js","vgo");
                   vgo('setAccount', '89584074');
                   vgo('setTrackByDefault', false);
                   vgo('process');
-        `}}/>
+        `,
+          }}
+        />
       </Html>
     )
   }
