+++ src/layout/ScreenSize.tsx
@@ -1,5 +1,4 @@
 import throttle from "lodash.throttle"
-import MobileDetect from "mobile-detect"
 import * as React from "react"
 import { Dimensions } from "react-native"
 import { BANNER_HEIGHT } from "src/header/BlueBanner"
@@ -112,29 +111,5 @@ export function useScreenSize() {
     isMobile: screen === ScreenSizes.MOBILE,
     isDesktop: screen === ScreenSizes.DESKTOP,
     isTablet: screen === ScreenSizes.TABLET,
-  }
-}
-
-// by guessing device type we can have our server rendered content likely be the right size
-// if not it will be fixed on the client after first rendeer.
-// note this is intended to be used serverside and only as a guess
-export function setDimensionsForScreen(userAgent: string) {
-  const md = new MobileDetect(userAgent)
-
-  if (md.mobile()) {
-    Dimensions.set({
-      window: { width: TABLET_BREAKPOINT - 1 },
-      screen: { width: TABLET_BREAKPOINT - 1 },
-    })
-  } else if (md.tablet()) {
-    Dimensions.set({
-      window: { width: DESKTOP_BREAKPOINT - 1 },
-      screen: { width: DESKTOP_BREAKPOINT - 1 },
-    })
-  } else {
-    Dimensions.set({
-      window: { width: DESKTOP_BREAKPOINT + 1 },
-      screen: { width: DESKTOP_BREAKPOINT + 1 },
-    })
   }
 }
